chore: refactor file names and locations
