+++ src/lib/staticValidationResults.ts
